@@ -20,68 +20,68 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^15.1.2",
-        "@angular/cdk": "^15.1.2",
-        "@angular/common": "^15.1.2",
-        "@angular/compiler": "^15.1.2",
-        "@angular/core": "^15.1.2",
-        "@angular/flex-layout": "^15.0.0-beta.42",
-        "@angular/forms": "^15.1.2",
-        "@angular/localize": "^15.1.2",
-        "@angular/material": "^15.1.2",
-        "@angular/platform-browser": "^15.1.2",
-        "@angular/platform-browser-dynamic": "^15.1.2",
-        "@angular/router": "^15.1.2",
-        "@apollo/client": "^3.7.5",
-        "@graphql-tools/mock": "^8.7.16",
-        "apollo-angular": "^4.2.0",
-        "graphql": "^15",
+        "@angular/animations": "^16.1.2",
+        "@angular/cdk": "^16.1.1",
+        "@angular/common": "^16.1.2",
+        "@angular/compiler": "^16.1.2",
+        "@angular/core": "^16.1.2",
+        "@angular/forms": "^16.1.2",
+        "@angular/localize": "^16.1.2",
+        "@angular/material": "^16.1.1",
+        "@angular/platform-browser": "^16.1.2",
+        "@angular/platform-browser-dynamic": "^16.1.2",
+        "@angular/router": "^16.1.2",
+        "@apollo/client": "^3.7.16",
+        "@graphql-tools/mock": "^9.0.0",
+        "@ngbracket/ngx-layout": "^16.0.0",
+        "apollo-angular": "^5.0.0",
+        "graphql": "^16.7.1",
         "lodash-es": "^4.17.21",
-        "prosemirror-commands": "^1.5.0",
-        "prosemirror-dropcursor": "^1.6.1",
-        "prosemirror-gapcursor": "^1.3.1",
-        "prosemirror-history": "^1.3.0",
-        "prosemirror-inputrules": "^1.2.0",
-        "prosemirror-keymap": "^1.2.0",
-        "prosemirror-menu": "^1.2.1",
-        "prosemirror-model": "^1.19.0",
-        "prosemirror-schema-basic": "^1.2.1",
-        "prosemirror-schema-list": "^1.2.2",
-        "prosemirror-state": "^1.4.2",
-        "prosemirror-tables": "^1.3.2",
-        "prosemirror-view": "^1.30.0",
-        "rxjs": "^7.8.0",
-        "tslib": "^2.5.0",
-        "zone.js": "~0.12.0"
+        "prosemirror-commands": "^1.5.2",
+        "prosemirror-dropcursor": "^1.8.1",
+        "prosemirror-gapcursor": "^1.3.2",
+        "prosemirror-history": "^1.3.2",
+        "prosemirror-inputrules": "^1.2.1",
+        "prosemirror-keymap": "^1.2.2",
+        "prosemirror-menu": "^1.2.2",
+        "prosemirror-model": "^1.19.2",
+        "prosemirror-schema-basic": "^1.2.2",
+        "prosemirror-schema-list": "^1.3.0",
+        "prosemirror-state": "^1.4.3",
+        "prosemirror-tables": "^1.3.4",
+        "prosemirror-view": "^1.31.5",
+        "rxjs": "^7.8.1",
+        "tslib": "^2.5.3",
+        "zone.js": "~0.13.1"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^15.1.3",
-        "@angular-eslint/builder": "15.2.0",
-        "@angular-eslint/eslint-plugin": "15.2.0",
-        "@angular-eslint/eslint-plugin-template": "15.2.0",
-        "@angular-eslint/schematics": "15.2.0",
-        "@angular-eslint/template-parser": "15.2.0",
-        "@angular/cli": "^15.1.3",
-        "@angular/compiler-cli": "^15.1.2",
-        "@angular/language-service": "^15.1.2",
-        "@playwright/test": "^1.30.0",
-        "@types/jasmine": "~4.3.1",
+        "@angular-devkit/build-angular": "^16.1.1",
+        "@angular-eslint/builder": "16.0.3",
+        "@angular-eslint/eslint-plugin": "16.0.3",
+        "@angular-eslint/eslint-plugin-template": "16.0.3",
+        "@angular-eslint/schematics": "16.0.3",
+        "@angular-eslint/template-parser": "16.0.3",
+        "@angular/cli": "^16.1.1",
+        "@angular/compiler-cli": "^16.1.2",
+        "@angular/language-service": "^16.1.2",
+        "@playwright/test": "^1.35.1",
+        "@types/jasmine": "~4.3.4",
         "@types/jasminewd2": "^2.0.10",
-        "@types/lodash-es": "^4.17.6",
+        "@types/lodash-es": "^4.17.7",
         "@types/node": "^16.0",
-        "@typescript-eslint/eslint-plugin": "^5.43.0",
-        "@typescript-eslint/parser": "^5.43.0",
-        "eslint": "^8.33.0",
-        "jasmine-core": "~4.5.0",
+        "@typescript-eslint/eslint-plugin": "^5.59.2",
+        "@typescript-eslint/parser": "^5.59.2",
+        "eslint": "^8.39.0",
+        "jasmine-core": "~5.0.1",
         "jasmine-spec-reporter": "~7.0.0",
-        "karma": "~6.4.1",
-        "karma-chrome-launcher": "~3.1.0",
+        "karma": "~6.4.2",
+        "karma-chrome-launcher": "~3.2.0",
         "karma-coverage": "^2.2.0",
         "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "^2.0.0",
-        "ng-packagr": "^15.1.1",
-        "prettier": "2.8.3",
+        "karma-jasmine-html-reporter": "^2.1.0",
+        "ng-packagr": "^16.1.0",
+        "prettier": "2.8.8",
         "ts-node": "^10.6.0",
-        "typescript": "~4.9.4"
+        "typescript": "~5.1.3"
     }
 }
