@@ -10,28 +10,26 @@
         }
     },
     "dependencies": {
-        "prosemirror-commands": "^1.1.12",
-        "prosemirror-dropcursor": "^1.4.0",
-        "prosemirror-gapcursor": "^1.2.1",
-        "prosemirror-history": "^1.2.0",
-        "prosemirror-inputrules": "^1.1.3",
-        "prosemirror-keymap": "^1.1.5",
-        "prosemirror-menu": "^1.1.4",
-        "prosemirror-model": "^1.16.1",
-        "prosemirror-schema-basic": "^1.1.2",
-        "prosemirror-schema-list": "^1.1.6",
-        "prosemirror-state": "^1.3.4",
-        "prosemirror-tables": "^1.1.1",
-        "prosemirror-view": "^1.23.5",
-        "tslib": "^2.3.0"
+        "prosemirror-commands": "^1.5.2",
+        "prosemirror-dropcursor": "^1.8.1",
+        "prosemirror-gapcursor": "^1.3.2",
+        "prosemirror-history": "^1.3.2",
+        "prosemirror-inputrules": "^1.2.1",
+        "prosemirror-keymap": "^1.2.2",
+        "prosemirror-menu": "^1.2.2",
+        "prosemirror-model": "^1.19.2",
+        "prosemirror-schema-basic": "^1.2.2",
+        "prosemirror-schema-list": "^1.3.0",
+        "prosemirror-state": "^1.4.3",
+        "prosemirror-tables": "^1.3.4",
+        "prosemirror-view": "^1.31.5",
+        "tslib": "^2.5.3"
     },
     "peerDependencies": {
-        "@angular/common": "^15.1.0",
-        "@angular/core": "^15.1.0",
-        "@angular/forms": "^15.1.0",
-        "@angular/material": "^15.1.0",
-        "@angular/platform-browser": "^15.1.0",
-        "@angular/platform-browser-dynamic": "^15.1.0",
-        "@ecodev/natural": ">40.0"
+        "@angular/common": "^16.1.0",
+        "@angular/core": "^16.1.0",
+        "@angular/forms": "^16.1.0",
+        "@angular/material": "^16.1.0",
+        "@ecodev/natural": ">50.0"
     }
 }
