@@ -10,22 +10,20 @@
         }
     },
     "dependencies": {
-        "tslib": "^2.3.1"
+        "tslib": "^2.5.3"
     },
     "peerDependencies": {
-        "@angular/animations": "^15.1.0",
-        "@angular/cdk": "^15.1.0",
-        "@angular/common": "^15.1.0",
-        "@angular/core": "^15.1.0",
-        "@angular/flex-layout": "^15.0.0-beta.42",
-        "@angular/forms": "^15.1.0",
-        "@angular/localize": "^15.1.0",
-        "@angular/material": "^15.1.0",
-        "@angular/platform-browser": "^15.1.0",
-        "@angular/platform-browser-dynamic": "^15.1.0",
-        "@angular/router": "^15.1.0",
-        "apollo-angular": "^4.2.0",
+        "@angular/animations": "^16.1.0",
+        "@angular/cdk": "^16.1.0",
+        "@angular/common": "^16.1.0",
+        "@angular/core": "^16.1.0",
+        "@angular/forms": "^16.1.0",
+        "@angular/localize": "^16.1.0",
+        "@angular/material": "^16.1.0",
+        "@angular/platform-browser": "^16.1.0",
+        "@angular/router": "^16.1.0",
+        "apollo-angular": "^5.0.0",
         "lodash-es": "^4.17.21",
-        "rxjs": "^7.4.0"
+        "rxjs": "^7.8.1"
     }
 }
