@@ -1,19 +1,12 @@
-import {Injectable} from '@angular/core';
-import {Resolve} from '@angular/router';
 import {Observable, of} from 'rxjs';
 import {Item, ItemService} from './item.service';
+import {inject} from '@angular/core';
 
-@Injectable({
-    providedIn: 'root',
-})
-export class AnyResolver implements Resolve<{model: Item}> {
-    public constructor(private readonly itemService: ItemService) {}
-
-    /**
-     * Resolve taxonomy data for router and panels service
-     */
-    public resolve(): Observable<{model: Item}> {
-        console.warn('resolve any');
-        return of({model: this.itemService.getItem(true)});
-    }
+/**
+ * Resolve taxonomy data for router and panels service
+ */
+export function resolveAny(): Observable<{model: Item}> {
+    const itemService = inject(ItemService);
+    console.warn('resolve any');
+    return of({model: itemService.getItem(true)});
 }
